# service/src/main/resources/api/service_openapi.yaml
openapi: 3.0.3
info:
  title: Tanagra Service API
  description: github.com/DataBiosphere/tanagra
  version: 0.0.0
servers:
  - url: /
security:
  - bearerAuth: []
  - authorization: [openid, email, profile]
paths:
...
  "/v2/studies/{studyId}":
    parameters:
      - $ref: '#/components/parameters/StudyIdV2'
    get:
      summary: Get an existing study
      operationId: getStudy
      tags: [StudiesV2]
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudyV2'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
    patch:
      summary: Update an existing study
      operationId: updateStudy
      tags: [StudiesV2]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudyUpdateInfoV2'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudyV2'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
    delete:
      summary: Delete a study
      operationId: deleteStudy
      tags: [StudiesV2]
      responses:
        204:
          description: OK
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
...
  "/v2/studies/{studyId}/conceptSets":
    parameters:
      - $ref: '#/components/parameters/StudyIdV2'
    get:
      parameters:
        - $ref: "#/components/parameters/OffsetV2"
        - $ref: "#/components/parameters/LimitV2"
      summary: List all concept sets in a study
      operationId: listConceptSets
      tags: [ConceptSetsV2]
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConceptSetListV2"
        500:
          $ref: "#/components/responses/ServerError"
    post:
      summary: Create a new concept set
      operationId: createConceptSet
      tags: [ConceptSetsV2]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConceptSetCreateInfoV2"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConceptSetV2"
        500:
          $ref: "#/components/responses/ServerError"
  "/v2/studies/{studyId}/conceptSets/{conceptSetId}":
    parameters:
      - $ref: '#/components/parameters/StudyIdV2'
      - $ref: '#/components/parameters/ConceptSetIdV2'
    get:
      summary: Get an existing concept set
      operationId: getConceptSet
      tags: [ConceptSetsV2]
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConceptSetV2'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
    patch:
      summary: Update an existing concept set
      operationId: updateConceptSet
      tags: [ConceptSetsV2]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConceptSetUpdateInfoV2'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConceptSetV2'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
    delete:
      summary: Delete a concept set
      operationId: deleteConceptSet
      tags: [ConceptSetsV2]
      responses:
        204:
          description: OK
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
...
components:
  parameters:
...
    StudyIdV2:
      name: studyId
      in: path
      description: ID of the study
      required: true
      schema:
        type: string
...
    ConceptSetIdV2:
      name: conceptSetId
      in: path
      description: ID of the concept set
      required: true
      schema:
        type: string
...
    OffsetV2:
      name: offset
      in: query
      description: The number of items to skip before starting to collect the result set.
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    LimitV2:
      name: limit
      in: query
      description: The maximum number of items to return. Default 15
      required: false
      schema:
        type: integer
        minimum: 1
        default: 15
  responses:
    NotFound:
      description: Not found (or unauthorized)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorReportV2'
    ServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorReportV2'
  schemas:
...
    StudyV2:
      type: object
      required: [id, created, createdBy, lastModified]
      properties:
        id:
          description: ID of the study, immutable
          type: string
        displayName:
          $ref: "#/components/schemas/StudyDisplayNameV2"
        description:
          $ref: "#/components/schemas/StudyDescriptionV2"
        properties:
          $ref: "#/components/schemas/PropertiesV2"
        created:
          type: string
          format: date-time
        createdBy:
          description: Email of user who created the study
          type: string
        lastModified:
          type: string
          format: date-time
...
    StudyUpdateInfoV2:
      type: object
      properties:
        displayName:
          $ref: "#/components/schemas/StudyDisplayNameV2"
        description:
          $ref: "#/components/schemas/StudyDescriptionV2"
    StudyDisplayNameV2:
      type: string
      description: Human readable name of the study
    StudyDescriptionV2:
      type: string
      description: Description of the study
    PropertiesV2:
      description: List of key-value pairs of strings
      type: array
      items:
        $ref: "#/components/schemas/PropertyKeyValueV2"

    # There seems to be a bug in the TS code generator that fails to create
    # proper objects for these when they're nested directy inside PropertiesV2.
    # It instead leaves them as untyped "object". My best guess is that it's
    # cause by having an array as the top level object and it fails to generate
    # a nested type for that case like for nested objects under other objects.
    # I've moved it out into a separate top level type as a workaround.
    PropertyKeyValueV2:
      type: object
      required: [ key, value ]
      properties:
        key:
          description: |
            String to identify the key. Keys must be 1-1024 characters, using letters,
            numbers, dashes, and underscores and must not start with a dash or underscore.
          type: string
        value:
          description: Arbitrary string value
          type: string
...
    ConceptSetV2:
      type: object
      properties:
        id:
          description: ID of the concept set, immutable
          type: string
        underlayName:
          $ref: "#/components/schemas/UnderlayNameV2"
        entity:
          $ref: "#/components/schemas/EntityNameV2"
        displayName:
          $ref: "#/components/schemas/ConceptSetDisplayNameV2"
        description:
          $ref: "#/components/schemas/ConceptSetDescriptionV2"
        criteria:
          $ref: "#/components/schemas/CriteriaV2"
        created:
          description: Timestamp of when the concept set was created
          type: string
          format: date-time
        createdBy:
          description: Email of user who created the concept set
          type: string
        lastModified:
          description: Timestamp of when the concept set was last modified
          type: string
          format: date-time
      required:
        - id
        - underlayName
        - entity
        - displayName
        - criteria
        - created
        - createdBy
        - lastModified
    ConceptSetListV2:
      type: array
      items:
        $ref: "#/components/schemas/ConceptSetV2"
    ConceptSetCreateInfoV2:
      type: object
      properties:
        underlayName:
          $ref: "#/components/schemas/UnderlayNameV2"
        entity:
          $ref: "#/components/schemas/EntityNameV2"
        displayName:
          $ref: "#/components/schemas/ConceptSetDisplayNameV2"
        description:
          $ref: "#/components/schemas/ConceptSetDescriptionV2"
        criteria:
          $ref: "#/components/schemas/CriteriaV2"
    ConceptSetUpdateInfoV2:
      type: object
      properties:
        entity:
          $ref: "#/components/schemas/EntityNameV2"
        displayName:
          $ref: "#/components/schemas/ConceptSetDisplayNameV2"
        description:
          $ref: "#/components/schemas/ConceptSetDescriptionV2"
        criteria:
          $ref: "#/components/schemas/CriteriaV2"
    EntityNameV2:
      type: string
      description: Name of the entity (e.g. condition_occurrence)
    ConceptSetDisplayNameV2:
      type: string
      description: Human readable name of the concept set
    ConceptSetDescriptionV2:
      type: string
      description: Description of the concept set
    UnderlayNameV2:
      type: string
      description: Name of the underlay, immutable
...
    CriteriaV2:
      type: object
      description: Single criteria for a cohort or concept set definition
      properties:
        id:
          type: string
          description: ID of the criteria, immutable
        displayName:
          type: string
          description: Name of the criteria
        pluginName:
          type: string
          description: Name of the plugin that generated this criteria
        selectionData:
          type: string # JSON formatted
          description: Serialized plugin-specific representation of the user's selection
        uiConfig:
          # TODO: Remove this uiConfig property from the service API once we fully support UI plugins, because it
          # should be identical to the static config stored in the underlay. For now, we save it separately for
          # each criteria so we can handle schema changes in the UI code instead of in plugin-specific backend code.
          type: string # JSON formatted
          description: Serialized plugin-specific UI configuration for the criteria
        tags:
          type: array
          description: Unstructured tags for the criteria
          items:
            type: string
      required:
        - id
        - displayName
        - pluginName
        - selectionData
        - uiConfig
...
    ErrorReportV2:
      type: object
      required: [message, statusCode, causes]
      properties:
        message:
          type: string
        statusCode:
          type: integer
        causes:
          type: array
          items:
            type: string
...
  securitySchemes:
    authorization:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://accounts.google.com/o/oauth2/auth
          scopes:
            openid: open id authorization
            email: email authorization
            profile: profile authorization
    bearerAuth:
      type: http
      scheme: bearer